modularized app but now it is not center
