+++ src/index.css
@@ -18,7 +18,7 @@
   -webkit-text-size-adjust: 100%;
 }
 body {
-  margin: 0;
+  margin: 0 auto;
   display: flex;
   place-items: center;
   min-width: 320px;
@@ -26,7 +26,7 @@ body {
 }
 
 /* not sure why this style is not applying */
-.img {
+/* .img {
   border: 1px solid #1AA2B0 !important;
   border-radius: 10px;
-}
+} */
